==> code/1. HS code translation.ipynb <==
HS CODE 10단위 영어 전처리

---
- Marian 모델로 성질통합분류코드명(한글) 영어로 번역
- 영문품목명 + 번역성질통합분류코드명(영어) = 품목명 컬럼 생성


## setting

import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

hs = pd.read_excel("/content/drive/MyDrive/관세청_HS부호_240101.xlsx")
hs

## preprocessing

hs['성질통합분류코드명'] = hs['성질통합분류코드명'].str.replace(r'[\(\)]', '', regex=True)
hs

## Marian 번역 모델 적용

from transformers import MarianMTModel, MarianTokenizer
import sentencepiece as spm

model_name = 'Helsinki-NLP/opus-mt-ko-en'
tokenizer = MarianTokenizer.from_pretrained(model_name)
model = MarianMTModel.from_pretrained(model_name)

def translate(text):
    if pd.isnull(text):
        return None
    inputs = tokenizer.encode(text, return_tensors='pt', max_length=512, truncation=True)
    translated = model.generate(inputs, max_length=512, num_beams=4, early_stopping=True)
    translated_text = tokenizer.decode(translated[0], skip_special_tokens=True)
    return translated_text

hs['성질통합분류코드명_영문'] = hs['성질통합분류코드명'].apply(translate)
hs

hs['품목명'] = hs['영문품목명'] + ' belonging to ' + hs['성질통합분류코드명_영문']
hs.to_json("HS10전처리.json", orient='records')

# 이후 수작업으로 전처리를 추가함
