==> code/3. Embedding.ipynb <==
## sentence transformer setting

pip install -U sentence-transformers

from sentence_transformers import SentenceTransformer
model = SentenceTransformer('sentence-transformers/all-mpnet-base-v2')

## google colab setting

from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import warnings
from pandas.errors import SettingWithCopyWarning
warnings.simplefilter(action='ignore', category=(SettingWithCopyWarning))

# data path
path = '/content/drive/MyDrive/python/공공데이터 공모전/data/'

## 1. DSC(import) embedding

#LLM의 수입 품목 데이터
data = pd.read_json(path+'LLMimport.json')

### preprocessing

# LLM result preprocessing
def preprocess(x):
  if ':' in x :
    x = x[x.find(':')+1:]
  # ":"가 있으면 그 전까지 문장 제거, '\n*' 는 ','으로 replace, "" 큰따옴표 제거
  x = x.replace('\n*', ',').replace('\n, ', '').replace('"', '').replace('\n', ',').strip()
  if x[:2] == ',,':
    x = x[2:]
  if ',,' in x:
    x = x.split(',,')[0] #상단 코드 \n replace 때문에 생긴 패턴으로, 뒷부분이 부연설명이므로 삭제
  # 문장 구분자 : 온점 붙이기
  if x[-1] != '.':
    x = x+'.'
  x = x.replace('...', '').replace('* ', '').replace(';', ', ').replace(',', ', ').replace('  ', ' ').strip()
  return x

#적용
data['LLMimport'] = data['LLMimport'].apply(lambda x : preprocess(x))

# 문장 형태로 기존의 DSC에 결합
data['DSC(import)'] = data['DSC'] + '. This company may need ' + data['LLMimport']

# ex (1)
data['DSC(import)'][32]

# ex (2)
data['DSC(import)'][10]

### embedding

data['transformer_embedding'] = '0' #object로 해줘야

for i in range(len(data)):
  text = data['DSC(import)'][i]
  try:
    vec = list(model.encode(text))
    data['transformer_embedding'][i] = vec
  except:
    pass

data.to_json(path+'data(import)_transformer_embedding.json')

data

## 2. HS10 전처리 embedding

# 수작업으로 전처리한 HS10 데이터
hs = pd.read_json(path+'HS10전처리.json')

### embedding

hs['transformer_embedding'] = '0'

for i in range(len(hs)):
  text = hs['품목명'][i]
  try:
    vec = list(model.encode(text))
    hs['transformer_embedding'][i] = vec
  except:
    pass

hs.to_json(path+'hs_transformer_embedding.json')

## 3. DSC 내의 except & other than embedding

data = pd.read_excel(path+'비식별된 해외기업별 영문 텍스트데이터.xlsx')

### preprocessing

# DSC 내에 except 토큰이 있는 경우만 필터링
def if_except(text):
  if 'except' in text:
    return 1
  else:
    return 0
data['except'] = data['DSC'].apply(lambda x : if_except(x))

# DSC 내에 other than 토큰이 있는 경우만 필터링
def if_other_than(text):
  if 'other than' in text:
    return 1
  else:
    return 0
data['other_than'] = data['DSC'].apply(lambda x : if_other_than(x))

#적용
data = data[(data['except']==1)|(data['other_than']==1)].drop_duplicates('DSC').reset_index(drop=True)
data = pd.DataFrame(data['DSC'])

# except에 해당하는 내용만 추출
def except_parentheses(x):
  if 'except' in x:
    x = x.replace(';', '.').replace(',', '.').replace(')', '.')
    x = x[x.find('except'):]
    x = x[:x.find('.')].replace('except ', '')
    return x
  else:
    return '0'

# other than에 해당하는 내용만 추출
def other_parentheses(x):
  if 'other than' in x:
    x = x.replace(';', '.').replace(',', '.').replace(')', '.')
    x = x[x.find('other than'):]
    x = x[:x.find('.')].replace('other than ', '')
    return x
  else:
    return '0'

#적용
data['except'] = data['DSC'].apply(lambda x : except_parentheses(x))
data['other'] = data['DSC'].apply(lambda x : other_parentheses(x))

### embedding

data['except_embedding'] = '0'
data['other_embedding'] = '0'

# except embedding
for i in list(data[data['except']!='0'].index):
  ex = data['except'][i]
  vec = list(model.encode(ex))
  data['except_embedding'][i] = vec

# other than embedding
for i in list(data[data['other']!='0'].index):
  ot = data['other'][i]
  vec = list(model.encode(ot))
  data['other_embedding'][i] = vec

data.to_json(path+'ex_other_transformer_embedding.json')
