==> code/5. Reranking.ipynb <==
## model setting

pip install -U FlagEmbedding

from FlagEmbedding import FlagReranker
reranker = FlagReranker('BAAI/bge-reranker-v2-m3', use_fp16=True)

## setting

import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
import warnings
from pandas.errors import SettingWithCopyWarning
warnings.simplefilter(action='ignore', category=(SettingWithCopyWarning))
import ast

from google.colab import drive
drive.mount('/content/drive')

#data path
path = '/content/drive/MyDrive/python/공공데이터 공모전/data/'
#load data
corp = pd.read_json(path+'mapping30.json', dtype = {'CODE':str})
hs_df = pd.read_json(path+'hs_transformer_embedding.json').explode('HS부호')

## preprocessing

def change_key_to_float(di):
  new = {}
  for key in di.keys():
    new[float(key)] = di[key]
  return new

corp['matched'] = corp['matched'].apply(lambda x : change_key_to_float(x))
corp

hs_dict = {}
for index, row in hs_df.iterrows():
    hs_dict[row["HS부호"]] = row["품목명"]

corp['cross_result'] = [{} for _ in range(len(corp))]

## reranking

for index, entry in corp.iterrows():
    dsc = entry["DSC"]
    matched = entry["matched"]

    pairs = []
    hs_codes_list = {}

    for hs_codes in matched.values():
        for hs_code in hs_codes:
            if hs_code in hs_dict:
                document = hs_dict[hs_code]
                pairs.append((dsc, document))
                if document not in hs_codes_list:
                    hs_codes_list[document] = []
                hs_codes_list[document].extend(hs_codes)
                break

    if pairs:

        scores = reranker.compute_score([(dsc, document) for dsc, document in pairs])

        cross_results = {}
        for (query, document), score in zip(pairs, scores):
            if score not in cross_results:
                cross_results[score] = []
            cross_results[score].append(hs_codes_list[document])

        sorted_cross_result = {k: cross_results[k] for k in sorted(cross_results, reverse=True)}
        corp.at[index, "cross_result"] = sorted_cross_result

# 정렬, 리스트 풀기
for index, entry in corp.iterrows():
    cross_result = entry["cross_result"]

    for score, hs_codes_list in cross_result.items():
        cross_result[score] = [item for sublist in hs_codes_list for item in sublist]

    sorted_cross_result = {k: cross_result[k] for k in sorted(cross_result, reverse=True)}
    corp.at[index, "cross_result"] = sorted_cross_result

corp.to_json('Reranker.json')
