==> code/7. EnsembleVoting Filtering Submission.ipynb <==
## setting

from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import warnings
from pandas.errors import SettingWithCopyWarning
warnings.simplefilter(action='ignore', category=(SettingWithCopyWarning))

path = '/content/drive/MyDrive/python/2024-1학기/공공데이터 공모전/data/'
data = pd.read_json(path + 'mapping30.json')
hs = pd.read_json(path+'HS10전처리.json').explode('HS부호').reset_index(drop=True)
ex_other = pd.read_json(path+'ex_other_mapping.json')
label = pd.read_json(path+'handlabel.json')
vae = pd.read_json(path+'VAE.json').drop_duplicates('DSC').reset_index(drop=True)
rerank = pd.read_json(path+'Reranker.json')
rerank['DSC'] = data['DSC']

for i in label[label['handlabel']=='none'].index:
  label['handlabel'][i] = []

ex_other = ex_other.drop(['except_embedding', 'other_embedding'], axis=1)
data = pd.merge(data, vae[['DSC', 'VAE_top_10']], on='DSC', how='left')
data = pd.merge(data, rerank[['DSC', 'cross_result']], on='DSC', how='left')
data = pd.merge(data, ex_other, on='DSC', how='left')
data = data.drop(['transformer_embedding', 'maxscore', 'ex_maxscore', 'ot_maxscore'], axis=1)
data = data.fillna('0')
data

def change_key_to_float(di):
  if di != '0':
    new = {}
    for key in di.keys():
      new[float(key)] = di[key]
    return new
  else:
    return '0'

data['matched'] = data['matched'].apply(lambda x : change_key_to_float(x))
data['VAE_top_10'] = data['VAE_top_10'].apply(lambda x : change_key_to_float(x))
data['cross_result'] = data['cross_result'].apply(lambda x : change_key_to_float(x))
data['ex_matched'] = data['ex_matched'].apply(lambda x : change_key_to_float(x))
data['ot_matched'] = data['ot_matched'].apply(lambda x : change_key_to_float(x))

def if_in_list(x, hslist):
  if x in hslist:
    return 1
  else:
    return 0

## function

# 각 모델의 결과를 voting
def ensemble_mapping_vote(data, row, cut1, cut2, cut3):
  matched = list(set(sum([data['matched'][row][x] for x in list(data['matched'][row].keys())[:cut1]], [])))
  vae_matched = list(set(sum([data['VAE_top_10'][row][x] for x in list(data['VAE_top_10'][row].keys())[:cut2]], [])))
  rerank_matched = list(set(sum([data['cross_result'][row][x] for x in list(data['cross_result'][row].keys())[:cut3]], [])))
  # 다수결 : 3개 중 2개 이상이 추천한 것, 추천 결과의 intersection에 있는 코드만 추출
  match1 = list(set(matched) & set(vae_matched))
  match2 = list(set(rerank_matched) & set(vae_matched))
  match3 = list(set(rerank_matched) & set(matched))
  hslist = list(set(match1 + match2 + match3))
  if len(hslist)==0:
    return '0'
  else:
    tmp = hs.copy()
    tmp['filter'] = tmp['HS부호'].apply(lambda x : if_in_list(x, hslist))
    tmp = tmp[tmp['filter']==1].reset_index(drop=True)
    total = list(tmp['HS부호'])
    return total

# except, other than에 해당하는 코드는 제거
def filter_except(data, row, cut1, cut2):
  hslist = list(set(data['total'][row]))
  if len(hslist)==0:
    return '0'
  else:
    tmp = hs.copy()
    tmp['filter'] = tmp['HS부호'].apply(lambda x : if_in_list(x, hslist))
    tmp = tmp[tmp['filter']==1].reset_index(drop=True)
    removelist = []
    try:
      exlist = list(set(sum([data['ex_matched'][row][x] for x in list(data['ex_matched'][row].keys())[:cut1]], [])))
      removelist = removelist + exlist
    except:
      pass
    try:
      otlist = list(set(sum([data['ot_matched'][row][x] for x in list(data['ot_matched'][row].keys())[:cut2]], [])))
      removelist = removelist + otlist
    except:
      pass
    try:
      filterlist = list(set(tmp['HS부호']) - set(removelist))
      tmp['filter'] = tmp['HS부호'].apply(lambda x : if_in_list(x, filterlist))
      tmp = tmp[tmp['filter']==1].reset_index(drop=True)
    except:
      pass
    total = list(tmp['HS부호'])
    return total

def acc_score(data, label):
  x = pd.merge(label, data[['DSC', 'DSC(import)', 'total']], on='DSC', how='left')
  x['inter'] = '0'
  x['len'] = 0
  x['score'] = 0
  for i in range(len(x)):
    x['inter'][i] = len(set(x['handlabel'][i])&set(x['total'][i]))
    x['len'][i] = len(x['total'][i])
    if x['len'][i] > 0:
      x['score'][i] = x['inter'][i]/x['len'][i]
    else:
      x['score'][i] = 'No match'
  return x

def show_total(data, row, show=False):
  dsc = data['DSC(import)'][row]
  hslist = data['total'][row]
  tmp = hs.copy()
  tmp['filter'] = tmp['HS부호'].apply(lambda x : if_in_list(x, hslist))
  tmp = tmp[tmp['filter']==1].reset_index(drop=True)
  if show==True:
    print('DSC(import) :', dsc)
    for text in list(set(tmp['품목명'])):
      print(text)
  return dsc, tmp

# 각 모델의 결과를 union
def ensemble_mapping_union(data, row, cut1=1, cut2=1, cut3=1):
  matched = list(set(sum([data['matched'][row][x] for x in list(data['matched'][row].keys())[:cut1]], [])))
  vae_matched = list(set(sum([data['VAE_top_10'][row][x] for x in list(data['VAE_top_10'][row].keys())[:cut2]], [])))
  rerank_matched = list(set(sum([data['cross_result'][row][x] for x in list(data['cross_result'][row].keys())[:cut3]], [])))
  # 각 모델이 추천한 top1을 union
  hslist = list(set(matched + vae_matched + rerank_matched))
  if len(hslist)==0:
    return '0'
  else:
    tmp = hs.copy()
    tmp['filter'] = tmp['HS부호'].apply(lambda x : if_in_list(x, hslist))
    tmp = tmp[tmp['filter']==1].reset_index(drop=True)
    total = list(tmp['HS부호'])
    return total

## ensemble voting

---

- 코사인 유사도 : top10
- Rerank : top10
- VAE : top5

선정기준 : handlabel score

result = data.copy()
result['total'] = '0'

# 앙상블 보팅
for i in range(len(result)):
  result['total'][i] = ensemble_mapping_vote(result, i, 10, 5, 10)

## filtering



---

- except : top10 제거
- other than : top10 제거

# 필터링
for i in range(len(result)):
  result['total'][i] = filter_except(result, i, 10, 10)

## Union

---

- 추천이 0개인 경우에 한하여 각 모델의 top1 추천을 union

result['len'] = result['total'].apply(lambda x : len(x))
result[result['len']==0]

for i in list(result[result['len']==0].index):
  result['total'][i] = ensemble_mapping_union(result, i, 1, 1, 1)

result['len'] = result['total'].apply(lambda x : len(x))
print(len(result[result['len']==0]))

## submission file

---

최종 결과물 excel file 생성

submission = pd.read_excel(path + '비식별된 해외기업별 영문 텍스트데이터.xlsx')
submission = pd.merge(submission, result[['DSC', 'total']], on='DSC', how='left')[['ID', 'total']]
submission = submission.explode('total').reset_index(drop=True)
print(sum(submission['total'].isnull()))
submission = submission.dropna().reset_index(drop=True)
submission

submission['total'] = submission['total'].astype(str)
def hscode(x):
  if len(x)==9:
    x = '0' + x
    return x
  else:
    return x
submission['total'] = submission['total'].apply(lambda x : hscode(x))

submission.to_excel(path+'submission.xlsx', index=False)

pd.read_excel(path+'submission.xlsx')
